修复代码告警问题

+++ services/camera_service/src/hcapture_session.cpp
@@ -1411,7 +1411,7 @@ bool StateMachine::Transfer(CaptureSessionState targetState)
 void HCaptureSession::UpdateCameraRotateAngleAndZoom(std::vector<CameraRotateStrategyInfo> &infos,
     std::vector<int32_t> &frameRateRange)
 {
-    int index = -1;
+    size_t index = -1;
     int uid = IPCSkeleton::GetCallingUid();
     std::string bundleName = GetClientBundle(uid);
     for (size_t i = 0; i < infos.size(); i++) {
